AK-47 balance
The 5-shot burst had too high accuracy. Bursts with rifles usually fire
3 shots with the same amount of accuracy as their lowest accuracy
singleshot. In the AK-47's case the lowest accuracy singleshot is at 40
(5 lower than default, but this is in compensation of the weapon's
mitigation of 5). A default burst with this weapon, 3 shots, would be
3x40 accuracy = 120., but since this weapon actually fires 5 shots, the
burst should be 5x24 accuracy = 120 (because 120/5=24).

Mathematically correct, and it's still good in game :) Close range is
where this weapon mostly should be used anyway, with this awesome
monster burst :)