Makarov tweak
With the new P7 changes, the Makarov needed some tweaking as to not get
left behind, so I lowered the TU cost a liiittle bit. Just a little
something to make it a bit more mobile, since it's supposed to be the
lightest pistol.

Also updated description in STRINGS.XML.